reorganises individual concerns into seperate modules, and tables into individual module files, corrects imports
