+++ be_parkfinite/main.py
@@ -1,10 +1,10 @@
 from fastapi import Depends, FastAPI, HTTPException
 from sqlalchemy.orm import Session
 
-import models, crud, schemas
-from database.database import SessionLocal, engine
+import models, be_parkfinite.crud.campsite_crud as campsite_crud, be_parkfinite.schemas.campsite_schemas as campsite_schemas
+from database.database import SessionLocal, engine, Base
 
-models.Base.metadata.create_all(bind=engine)
+Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
 
@@ -20,11 +20,11 @@ def root():
     return {"Hello": "World"}
 
 
-@app.get("/campsites", response_model=list[schemas.Campsite])
+@app.get("/campsites", response_model=list[campsite_schemas.Campsite])
 def read_campsites(skip: int = 0, limit: int = 5, db: Session = Depends(get_db)):
-    campsites = crud.get_campsites(db, skip=skip, limit=limit)
+    campsites = campsite_crud.get_campsites(db, skip=skip, limit=limit)
     return campsites
 
-@app.post("/campsites", response_model=schemas.Campsite)
-def create_campsite(campsite: schemas.CampsiteCreate, db: Session = Depends(get_db)):
-    return crud.create_campsite(db=db, campsite=campsite)
+@app.post("/campsites", response_model=campsite_schemas.Campsite)
+def create_campsite(campsite: campsite_schemas.CampsiteCreate, db: Session = Depends(get_db)):
+    return campsite_crud.create_campsite(db=db, campsite=campsite)
