+++ be_parkfinite/schemas/campsite_schemas.py
@@ -1,49 +1,5 @@
 from pydantic import BaseModel
 
-class CampsiteActivityBase(BaseModel):
-    campsite_id: int
-    activity_id: int
-
-class CampsiteActivityCreate(CampsiteActivityBase):
-    pass
-
-class CampsiteActivity(CampsiteActivityBase):
-    campsite_activity_id: int
-
-
-class ActivityBase(BaseModel):
-    activity_name: str
-    activity_img_url: str
-
-class ActivityCreate(ActivityBase):
-    pass
-
-class Activity(ActivityBase):
-    activity_id: int
-
-
-class CampsiteFacilityBase(BaseModel):
-    campsite_id: int
-    facility_id: int
-
-class CampsiteFacilityCreate(CampsiteFacilityBase):
-    pass
-
-class CampsiteFacility(CampsiteFacilityBase):
-    campsite_facility_id: int
-
-
-class FacilityBase(BaseModel):
-    facility_name: str
-    facility_img_url: str
-
-class FacilityCreate(FacilityBase):
-    pass
-
-class Facility(FacilityBase):
-    facility_id: int
-   
-
 class CampsitePhotoBase(BaseModel):
     campsite_photo_url: str
